requirements.txt
ultralytics==8.3.141
supervision==0.25.1
opencv-python==4.11.0.86
fastapi==0.115.12
uvicorn==0.34.2
pydantic==2.11.4
pydantic-settings==2.9.1
loguru==0.7.3
python-multipart==0.0.20
pillow==11.2.1